+++ install.sh
@@ -3,27 +3,29 @@
 set -e
 set -o pipefail
 
+error() {
+  echo "ERROR: $@" 1>&2
+  echo "Exiting installer" 1>&2
+  exit 1
+}
+
 ARCH=`uname -m`
 
 if [ "${ARCH}" != "x86_64" ]; then
-  echo "summon only works on 64-bit systems"
-  echo "exiting installer"
-  exit 1
+  error "summon only works on 64-bit systems"
 fi
 
 DISTRO=`uname | tr "[:upper:]" "[:lower:]"`
 
 if [ "${DISTRO}" != "linux" ] && [ "${DISTRO}" != "darwin"  ]; then
-  echo "This installer only supports Linux and OSX"
-  echo "exiting installer"
-  exit 1
+  error "This installer only supports Linux and OSX"
 fi
 
+tmp="/tmp"
 if [ ! -z "$TMPDIR" ]; then
-  tmp="/tmp"
-else
   tmp=$TMPDIR
 fi
+
 # secure-ish temp dir creation without having mktemp available (DDoS-able but not exploitable)
 tmp_dir="$tmp/install.sh.$$"
 (umask 077 && mkdir $tmp_dir) || exit 1
@@ -37,7 +39,6 @@ do_download() {
     curl --fail -sSL -o "$2" "$1" &>/dev/null || true
   else
     error "Could not find wget or curl"
-    exit 1
   fi
 }
 
@@ -52,9 +53,8 @@ get_latest_version() {
     latest_payload=$(curl --fail -sSL "$LATEST_VERSION_URL")
   else
     error "Could not find wget or curl"
-    exit 1
   fi
-  
+
   echo "$latest_payload" | # Get latest release from GitHub api
     grep '"tag_name":' | # Get tag line
     sed -E 's/.*"([^"]+)".*/\1/' # Pluck JSON value
@@ -62,7 +62,7 @@ get_latest_version() {
 
 LATEST_VERSION=$(get_latest_version)
 
-echo "Using version number: v$LATEST_VERSION"
+echo "Using version number: $LATEST_VERSION"
 
 BASEURL="https://github.com/cyberark/summon/releases/download/"
 URL=${BASEURL}"${LATEST_VERSION}/summon-${DISTRO}-amd64.tar.gz"
@@ -70,12 +70,12 @@ URL=${BASEURL}"${LATEST_VERSION}/summon-${DISTRO}-amd64.tar.gz"
 ZIP_PATH="${tmp_dir}/summon.tar.gz"
 do_download ${URL} ${ZIP_PATH}
 
-echo "Installing summon v${LATEST_VERSION} into /usr/local/bin"
+echo "Installing summon ${LATEST_VERSION} into /usr/local/bin"
 
 if sudo -h >/dev/null 2>&1; then
-  sudo tar -C /usr/local/bin -zxvf ${ZIP_PATH} >/dev/null
+  sudo tar -C /usr/local/bin -o -zxvf ${ZIP_PATH} >/dev/null
 else
-  tar -C /usr/local/bin -zxvf ${ZIP_PATH} >/dev/null
+  tar -C /usr/local/bin -o -zxvf ${ZIP_PATH} >/dev/null
 fi
 
 if [ -d "/etc/bash_completion.d" ]; then